# health_pal_backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  api_types.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  api_types.SuccessResponse:
    properties:
      message:
        type: string
    type: object
...
  handlers.RecordStepsRequest:
    properties:
      date:
        type: string
      steps_count:
        minimum: 0
        type: integer
    required:
    - date
    - steps_count
    type: object
  handlers.RecordWaterIntakeRequest:
    properties:
      amount_ml:
        minimum: 0
        type: integer
      date:
        type: string
    required:
    - amount_ml
    - date
    type: object
...
  models.StepResponse:
    properties:
      created_at:
        type: string
      date:
        type: string
      id:
        type: integer
      steps_count:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  models.WaterIntakeResponse:
    properties:
      amount_ml:
        type: integer
      created_at:
        type: string
      date:
        type: string
      id:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API documentation for the Health Pal backend.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Health Pal Backend API
  version: "1.0"
paths:
...
  /api/v1/steps:
    post:
      consumes:
      - application/json
      description: Records or updates the daily step count for the authenticated user.
      parameters:
      - description: Steps data
        in: body
        name: steps
        required: true
        schema:
          $ref: '#/definitions/handlers.RecordStepsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Steps recorded successfully
          schema:
            $ref: '#/definitions/api_types.SuccessResponse'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/api_types.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api_types.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Record daily steps
      tags:
      - Steps
  /api/v1/steps/history:
    get:
      description: Retrieves all recorded step counts for the authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: List of step entries
          schema:
            items:
              $ref: '#/definitions/models.StepResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api_types.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get step history
      tags:
      - Steps
  /api/v1/water-intakes:
    post:
      consumes:
      - application/json
      description: Records or updates the daily water intake in milliliters for the
        authenticated user.
      parameters:
      - description: Water intake data
        in: body
        name: waterIntake
        required: true
        schema:
          $ref: '#/definitions/handlers.RecordWaterIntakeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Water intake recorded successfully
          schema:
            $ref: '#/definitions/api_types.SuccessResponse'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/api_types.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api_types.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Record daily water intake
      tags:
      - WaterIntake
  /api/v1/water-intakes/history:
    get:
      description: Retrieves all recorded water intake in milliliters for the authenticated
        user.
      produces:
      - application/json
      responses:
        "200":
          description: List of water intake entries
          schema:
            items:
              $ref: '#/definitions/models.WaterIntakeResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api_types.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get water intake history
      tags:
      - WaterIntake
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"